Main commit to remove access token - #216
