+++ bench/gl-stats.html
@@ -12,7 +12,6 @@ window.performance.now = () => now;
 </head>
 <body style="margin: 0; padding: 0"><div id="map" style="width: 500px; height: 500px"></div></body>
 <script>
-maplibregl.accessToken = 'MAPBOX_ACCESS_TOKEN';
 maplibregl.workerCount = 1;
 
 var map = new maplibregl.Map({
