+++ bench/styles/benchmarks.js
@@ -1,5 +1,4 @@
 import maplibregl from '../../src';
-import accessToken from '../lib/access_token';
 import locationsWithTileID from '../lib/locations_with_tile_id';
 import styleBenchmarkLocations from '@mapbox/gazetteer/benchmark/style-benchmark-locations.json';
 import StyleLayerCreate from '../benchmarks/style_layer_create';
@@ -12,8 +11,6 @@ import QueryBox from '../benchmarks/query_box';
 import getWorkerPool from '../../src/util/global_worker_pool';
 
 const locations = locationsWithTileID(styleBenchmarkLocations.features);
-
-maplibregl.accessToken = accessToken;
 
 const benchmarks = window.benchmarks = [];
