+++ debug/iframe-blob.html
@@ -36,7 +36,6 @@ document.querySelector('iframe').src = URL.createObjectURL(new Blob([ `
 
 <script src='${js.href}'><\/script>
 <script>
-maplibregl.accessToken = '${maplibregl.accessToken}';
 var map = window.map = new maplibregl.Map({
     container: 'map',
     zoom: 12.5,
