--- debug/token.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>MapLibre GL JS debug page</title>
-    <meta charset='utf-8'>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel='stylesheet' href='../dist/maplibre-gl.css' />
-    <style>
-        body { margin: 0; padding: 0; }
-        html, body, #map { height: 100%; }
-    </style>
-</head>
-
-<body>
-<div id='map'></div>
-<div id='hidden' style='display:none'></div>
-<script src='../dist/maplibre-gl-dev.js'></script>
-<script>
-    // 1) first map should work with its valid local accessToken, despite the bad global mapboxgl token
-    // 2) first map should keep working properly, despite the bad token in the second map
-
-    var mapToken = maplibregl.accessToken;
-    maplibregl.accessToken = 'notAToken';
-
-    new maplibregl.Map({
-        container: 'map',
-        zoom: 12.5,
-        center: [-77.01866, 38.888],
-        style: 'https://demotiles.maplibre.org/style.json',
-        hash: true,
-        accessToken: mapToken
-    });
-
-    // var map2 = new maplibregl.Map({
-    //     container: 'hidden',
-    //     zoom: 12.5,
-    //     center: [-77.01866, 38.888],
-    //     style: 'https://demotiles.maplibre.org/style.json',
-    //     hash: true,
-    //     accessToken: 'anotherBadToken'
-    // });
-
-</script>
-</body>
-</html>
