+++ src/ui/control/attribution_control.ts
@@ -39,7 +39,6 @@ class AttributionControl implements IControl {
 
         bindAll([
             '_toggleAttribution',
-            '_updateEditLink',
             '_updateData',
             '_updateCompact'
         ], this);
@@ -66,11 +65,9 @@ class AttributionControl implements IControl {
         }
 
         this._updateAttributions();
-        this._updateEditLink();
 
         this._map.on('styledata', this._updateData);
         this._map.on('sourcedata', this._updateData);
-        this._map.on('moveend', this._updateEditLink);
 
         if (compact === undefined) {
             this._map.on('resize', this._updateCompact);
@@ -85,7 +82,6 @@ class AttributionControl implements IControl {
 
         this._map.off('styledata', this._updateData);
         this._map.off('sourcedata', this._updateData);
-        this._map.off('moveend', this._updateEditLink);
         this._map.off('resize', this._updateCompact);
 
         this._map = undefined;
@@ -108,35 +104,9 @@ class AttributionControl implements IControl {
         }
     }
 
-    _updateEditLink() {
-        let editLink = this._editLink;
-        if (!editLink) {
-            editLink = this._editLink = (this._container.querySelector('.mapbox-improve-map') as any);
-        }
-
-        const params = [
-            {key: 'owner', value: this.styleOwner},
-            {key: 'id', value: this.styleId},
-            {key: 'access_token', value: this._map._requestManager._customAccessToken || config.ACCESS_TOKEN}
-        ];
-
-        if (editLink) {
-            const paramString = params.reduce((acc, next, i) => {
-                if (next.value) {
-                    acc += `${next.key}=${next.value}${i < params.length - 1 ? '&' : ''}`;
-                }
-                return acc;
-            }, '?');
-            editLink.href = `${config.FEEDBACK_URL}/${paramString}${this._map._hash ? this._map._hash.getHashString(true) : ''}`;
-            editLink.rel = 'noopener nofollow';
-            this._setElementTitle(editLink, 'MapFeedback');
-        }
-    }
-
     _updateData(e: any) {
         if (e && (e.sourceDataType === 'metadata' || e.sourceDataType === 'visibility' || e.dataType === 'style')) {
             this._updateAttributions();
-            this._updateEditLink();
         }
     }
