+++ test/release/index.html
@@ -9,7 +9,6 @@
     <script>
         const mapboxgl = {};
     </script>
-    <script src='../../debug/access_token_generated.js'></script>
     <script src="index.js"></script>
 </head>
