+++ test/unit/util/ajax.test.js
@@ -78,18 +78,6 @@ test('ajax', (t) => {
         window.server.respond();
     });
 
-    t.test('getJSON, 401: Mapbox domain', (t) => {
-        window.server.respondWith(request => {
-            request.respond(401);
-        });
-        getJSON({url:'api.mapbox.com'}, (error) => {
-            t.equal(error.status, 401);
-            t.equal(error.message, "Unauthorized: you may have provided an invalid Mapbox access token. See https://www.mapbox.com/api-documentation/#access-tokens-and-token-scopes");
-            t.end();
-        });
-        window.server.respond();
-    });
-
     t.test('postData, 204(no content): no error', (t) => {
         window.server.respondWith(request => {
             request.respond(204);
